MAJOR Fixes, also Unicode Passwords
So it turns out that due to an error in how stuff was passed between the code that did passwords, regardless of what was entered in the password box, the key would end up being the same thing. In trying to fix it, I ended up moving to a different SHA256 implementation out of process-of-elimination. This one doesn't need me to do a hex-to-uint8array conversion of the generated key to fit the AES implementation's desire for a Uint8array input. The implementation I used is under MIT License (so still safe), and it has a feature the previous one didn't have, namely the ability for the passwords to be Unicode. This should make it easier to use the program as input into itself for multi-layer security. This fix took more-or-less two days, and the answer stared me in the face. This program is still compatible with earlier files that were either unencrypted (left on default key), or any old file if uou don't touch the box. Unicode passwords won't work with older versions, but those versions are more-or-less unencrypted no matter what is fed to them. This whole bug did not used to exist when I first made the code and for a while after, but older browsers and versions of the program STILL produced it, so this inexplicable bug HAD to be fixed. It worked. This change also fixed another terrible bug in the same area. And I fixed another trouble spot!